+++ coverage.js
@@ -1,0 +1,4 @@
+import jsdomify from 'jsdomify';
+jsdomify.create();
+
+require('./test/blog-post');
